+++ conftest.py
@@ -6,7 +6,8 @@ from django.utils.timezone import now
 from oauth2_provider.models import Application
 from rest_framework.test import APIClient
 
-from users.factories import FCMDeviceFactory, RecoveryStatusFactory, UserFactory
+from users.auth import SessionTokenAuthentication
+from users.factories import ConfigurationSessionFactory, FCMDeviceFactory, RecoveryStatusFactory, UserFactory
 
 
 @pytest.fixture
@@ -59,6 +60,12 @@ def authed_client(api_client, oauth_app):
 
 
 @pytest.fixture
+def authed_client_token(authed_client, valid_token):
+    authed_client.defaults["HTTP_AUTHORIZATION"] = f"Bearer {valid_token.key}"
+    return authed_client
+
+
+@pytest.fixture
 def recovery_status():
     return RecoveryStatusFactory()
 
@@ -69,3 +76,19 @@ def recovery_status_with_expired_token_user():
     status.user.deactivation_token_valid_until = now() - timedelta(days=1)
     status.user.save()
     return status
+
+
+@pytest.fixture
+def token_auth():
+    return SessionTokenAuthentication()
+
+
+@pytest.fixture
+def valid_token():
+    return ConfigurationSessionFactory()
+
+
+@pytest.fixture
+def expired_token():
+    expires = now() - timedelta(days=2)
+    return ConfigurationSessionFactory(expires=expires)
